feat: update People model and create people service validations

+++ src/services/people.service.js
@@ -1,4 +1,25 @@
 const People = require('../database/models/People');
+const Joi = require('joi');
+const errorMessage = require('../utils/errorMessageConstructor');
+const verifyId = require('../utils/verifyId');
+
+const PersonSchema = Joi.object({
+  name: Joi.string().required(),
+  lastName: Joi.string().required(),
+  email: Joi.string().email().required(),
+  role: Joi.string().required(),
+  city: Joi.string(),
+  age: Joi.number(),
+});
+
+const create = async (person) => {
+  const { error } = PersonSchema.validate(person);
+  if (error) throw errorMessage(400, error.message);
+
+  const newPerson = await People.create(person);
+
+  return newPerson;
+}
 
 const getAll = async () => {
   const people = await People.find({});
@@ -7,23 +28,28 @@ const getAll = async () => {
 };
 
 const getById = async (id) => {
+  const idIsValid = verifyId(id);
+  if (!idIsValid) throw errorMessage(400, 'Id is not in a valid format');
+
   const person = await People.findById(id);
+  if (person === null) throw errorMessage(404, 'Id is not found');
 
   return person;
 }
 
-const create = async (person) => {
-  const newPerson = await People.create(person);
-
-  return newPerson;
-}
-
 const deleteById = (id) => {
+  const idIsValid = verifyId(id);
+  if (!idIsValid) throw errorMessage(400, 'Id is not in a valid format');
+
   return People.findByIdAndDelete(id);
 }
 
 const updateById = async (id, person) => {
-  const updatedPerson = await People.findByIdAndUpdate(id, person);
+  const idIsValid = verifyId(id);
+  if (!idIsValid) throw errorMessage(400, 'Id is not in a valid format');
+
+  const updatedPerson = await People.findByIdAndUpdate(id, person, { new: true });
+  if (updatedPerson === null) throw errorMessage(404, 'Id is not found');
 
   return updatedPerson;
 };
